postgres/requirements.txt
Faker==37.12.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
setuptools==80.9.0
tzdata==2025.2
wheel==0.45.1
